requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1

# LangChain and LangGraph
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.20
langchain-community==0.0.13

# API clients
openai==1.6.1
replicate>=1.0.7
elevenlabs==0.2.27
tavily-python==0.3.0
google-generativeai>=0.3.0

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Video processing
ffmpeg-python==0.2.0
moviepy==1.0.3

# Audio processing
pydub==0.25.1

# HTTP and utilities
requests==2.31.0
aiohttp==3.9.1

# Logging
structlog==23.2.0
colorama==0.4.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Type checking
mypy==1.7.1
types-requests==2.31.0.10
types-PyYAML==6.0.12.12
